Refactor canonicalization into its own reduction
