+++ cvxpy/tests/test_qp.py
@@ -20,7 +20,6 @@ along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 import numpy
 from scipy.linalg import lstsq
 
-import cvxpy
 from cvxpy import Minimize, Problem
 from cvxpy.atoms import QuadForm, abs, power, quad_over_lin, sum_entries, sum_squares
 from cvxpy.expressions.variables import Variable
